+++ SharePointFramework/ProjectExtensions/src/components/DocumentTemplateDialog/index.tsx
@@ -1,33 +1,73 @@
-import { FileAddResult } from '@pnp/sp';
-import { DefaultButton, PrimaryButton } from 'office-ui-fabric-react/lib/Button';
-import { Selection } from 'office-ui-fabric-react/lib/DetailsList';
-import { MessageBarType } from 'office-ui-fabric-react/lib/MessageBar';
-import { ProgressIndicator } from 'office-ui-fabric-react/lib/ProgressIndicator';
-import * as strings from 'ProjectExtensionsStrings';
-import * as React from 'react';
-import * as formatString from 'string-format';
-import { SPDataAdapter } from '../../data';
-import { TemplateFile } from '../../models/index';
-import { BaseDialog } from '../@BaseDialog/index';
-import { InfoMessage } from '../InfoMessage';
-import styles from './DocumentTemplateDialog.module.scss';
-import { DocumentTemplateDialogScreen } from './DocumentTemplateDialogScreen';
-import { DocumentTemplateDialogScreenEditCopy } from './DocumentTemplateDialogScreenEditCopy';
-import { DocumentTemplateDialogScreenSelect } from './DocumentTemplateDialogScreenSelect';
-import { IDocumentTemplateDialogProps } from './IDocumentTemplateDialogProps';
+import { FileAddResult } from '@pnp/sp'
+import { DefaultButton, PrimaryButton } from 'office-ui-fabric-react/lib/Button'
+import { Selection } from 'office-ui-fabric-react/lib/DetailsList'
+import { MessageBarType } from 'office-ui-fabric-react/lib/MessageBar'
+import { ProgressIndicator } from 'office-ui-fabric-react/lib/ProgressIndicator'
+import * as strings from 'ProjectExtensionsStrings'
+import * as React from 'react'
+import * as formatString from 'string-format'
+import { SPDataAdapter } from '../../data'
+import { TemplateFile } from '../../models/index'
+import { BaseDialog } from '../@BaseDialog/index'
+import { InfoMessage } from '../InfoMessage'
+import styles from './DocumentTemplateDialog.module.scss'
+import { DocumentTemplateDialogScreen } from './DocumentTemplateDialogScreen'
+import { DocumentTemplateDialogScreenEditCopy } from './DocumentTemplateDialogScreenEditCopy'
+import { DocumentTemplateDialogScreenSelect } from './DocumentTemplateDialogScreenSelect'
+import { IDocumentTemplateDialogProps } from './IDocumentTemplateDialogProps'
 
 // tslint:disable-next-line: naming-convention
 export const DocumentTemplateDialog = (props: IDocumentTemplateDialogProps) => {
-    const selection = new Selection({ onSelectionChanged });
+    // eslint-disable-next-line @typescript-eslint/no-use-before-define
+    const selection = new Selection({ onSelectionChanged })
 
-    let [selected, setSelected] = React.useState([]);
-    const [progress, setProgress] = React.useState(null);
-    const [screen, setScreen] = React.useState(DocumentTemplateDialogScreen.Select);
-    const [isBlocking, setIsBlocking] = React.useState(false);
-    const [uploaded, setUploaded] = React.useState([]);
+    const [selected, setSelected] = React.useState([])
+    const [progress, setProgress] = React.useState(null)
+    const [screen, setScreen] = React.useState(DocumentTemplateDialogScreen.Select)
+    const [isBlocking, setIsBlocking] = React.useState(false)
+    const [uploaded, setUploaded] = React.useState([])
+
+    function onSelectionChanged() {
+        setSelected(selection.getSelection() as TemplateFile[])
+    }
+
+    /**
+     * On copy documents to web
+     * 
+     * @param {TemplateFile[]} templates Templates
+     * @param {string} folderServerRelativeUrl Folder URL
+     */
+    async function onStartCopy(templates: TemplateFile[], folderServerRelativeUrl: string): Promise<void> {
+        setScreen(DocumentTemplateDialogScreen.CopyProgress)
+        setIsBlocking(true)
+
+        const folder = SPDataAdapter.sp.web.getFolderByServerRelativeUrl(folderServerRelativeUrl)
+        const filesAdded: FileAddResult[] = []
+
+        for (let i = 0; i < templates.length; i++) {
+            const template = templates[i]
+            setProgress({ description: template.newName, percentComplete: (i / templates.length) })
+            try {
+                filesAdded.push(await template.copyTo(folder))
+            } catch (error) { }
+        }
+
+        selection.setItems([], true)
+        setScreen(DocumentTemplateDialogScreen.Summary)
+        setUploaded(filesAdded)
+        setIsBlocking(false)
+    }
+
+    /**
+     * On close dialog
+     */
+    function onClose() {
+        props.onDismiss({ reload: screen === DocumentTemplateDialogScreen.Summary })
+    }
 
 
     function onRenderContent() {
+        // eslint-disable-next-line default-case
         switch (screen) {
             case DocumentTemplateDialogScreen.Select: {
                 return (
@@ -36,7 +76,7 @@ export const DocumentTemplateDialog = (props: IDocumentTemplateDialogProps) => {
                         selection={selection}
                         selectedItems={selected}
                         templateLibrary={props.templateLibrary} />
-                );
+                )
             }
             case DocumentTemplateDialogScreen.EditCopy: {
                 return (
@@ -45,13 +85,13 @@ export const DocumentTemplateDialog = (props: IDocumentTemplateDialogProps) => {
                         libraries={props.libraries}
                         onStartCopy={onStartCopy}
                         onChangeScreen={s => setScreen(s)} />
-                );
+                )
             }
             case DocumentTemplateDialogScreen.CopyProgress: {
-                return <ProgressIndicator label={strings.CopyProgressLabel} {...progress} />;
+                return <ProgressIndicator label={strings.CopyProgressLabel} {...progress} />
             }
             case DocumentTemplateDialogScreen.Summary: {
-                return <InfoMessage type={MessageBarType.success} text={formatString(strings.SummaryText, uploaded.length)} />;
+                return <InfoMessage type={MessageBarType.success} text={formatString(strings.SummaryText, uploaded.length)} />
             }
         }
     }
@@ -66,60 +106,21 @@ export const DocumentTemplateDialog = (props: IDocumentTemplateDialogProps) => {
                             onClick={() => setScreen(DocumentTemplateDialogScreen.EditCopy)}
                             disabled={selected.length === 0} />
                     </>
-                );
+                )
             }
             case DocumentTemplateDialogScreen.Summary: {
                 return (
                     <>
-                        <DefaultButton text={strings.GetMoreText} onClick={_ => setScreen(DocumentTemplateDialogScreen.Select)} />
+                        <DefaultButton text={strings.GetMoreText} onClick={() => setScreen(DocumentTemplateDialogScreen.Select)} />
                         <DefaultButton text={strings.CloseModalText} onClick={onClose} />
                     </>
-                );
+                )
             }
             default: {
-                return null;
+                return null
             }
         }
     }
-
-    function onSelectionChanged() {
-        setSelected(selection.getSelection() as TemplateFile[]);
-    }
-
-    /**
-     * On copy documents to web
-     * 
-     * @param {TemplateFile[]} templates Templates
-     * @param {string} folderServerRelativeUrl Folder URL
-     */
-    async function onStartCopy(templates: TemplateFile[], folderServerRelativeUrl: string): Promise<void> {
-        setScreen(DocumentTemplateDialogScreen.CopyProgress);
-        setIsBlocking(true);
-
-        let folder = SPDataAdapter.sp.web.getFolderByServerRelativeUrl(folderServerRelativeUrl);
-        let filesAdded: FileAddResult[] = [];
-
-        for (let i = 0; i < templates.length; i++) {
-            const template = templates[i];
-            setProgress({ description: template.newName, percentComplete: (i / templates.length) });
-            try {
-                filesAdded.push(await template.copyTo(folder));
-            } catch (error) { }
-        }
-
-        selection.setItems([], true);
-        setScreen(DocumentTemplateDialogScreen.Summary);
-        setUploaded(filesAdded);
-        setIsBlocking(false);
-    }
-
-    /**
-     * On close dialog
-     */
-    function onClose() {
-        props.onDismiss({ reload: screen === DocumentTemplateDialogScreen.Summary });
-    }
-
 
     return (
         <BaseDialog
@@ -130,9 +131,9 @@ export const DocumentTemplateDialog = (props: IDocumentTemplateDialogProps) => {
             containerClassName={styles.documentTemplateDialog}>
             {onRenderContent()}
         </BaseDialog>
-    );
-};
+    )
+}
 
-export * from './IDocumentTemplateDialogDismissProps';
-export { IDocumentTemplateDialogProps };
+export * from './IDocumentTemplateDialogDismissProps'
+export { IDocumentTemplateDialogProps }