+++ SharePointFramework/ProjectExtensions/src/components/TemplateSelectDialog/ExtensionsSection/index.tsx
@@ -1,9 +1,9 @@
-import { stringIsNullOrEmpty } from '@pnp/common';
-import { Toggle } from 'office-ui-fabric-react/lib/Toggle';
-import * as React from 'react';
-import { ProjectTemplate } from '../../../models';
-import styles from './ExtensionsSection.module.scss';
-import { IExtensionsSectionProps } from './IExtensionsSectionProps';
+import { stringIsNullOrEmpty } from '@pnp/common'
+import { Toggle } from 'office-ui-fabric-react/lib/Toggle'
+import * as React from 'react'
+import { ProjectTemplate } from '../../../models'
+import styles from './ExtensionsSection.module.scss'
+import { IExtensionsSectionProps } from './IExtensionsSectionProps'
 
 // tslint:disable-next-line: naming-convention
 export const ExtensionsSection = (props: IExtensionsSectionProps) => {
@@ -14,13 +14,13 @@ export const ExtensionsSection = (props: IExtensionsSectionProps) => {
      * @param {boolean} checked Checked
      */
     const onChange = (extension: ProjectTemplate, checked: boolean): void => {
-        let selectedExtensions = [];
-        if (checked) selectedExtensions = [extension, ...props.selectedExtensions];
-        else selectedExtensions = props.selectedExtensions.filter(ext => extension.text !== ext.text);
-        props.onChange(selectedExtensions);
-    };
+        let selectedExtensions = []
+        if (checked) selectedExtensions = [extension, ...props.selectedExtensions]
+        else selectedExtensions = props.selectedExtensions.filter(ext => extension.text !== ext.text)
+        props.onChange(selectedExtensions)
+    }
 
-    const selectedKeys = props.selectedExtensions.map(lc => lc.key);
+    const selectedKeys = props.selectedExtensions.map(lc => lc.key)
 
 
     return (
@@ -40,5 +40,5 @@ export const ExtensionsSection = (props: IExtensionsSectionProps) => {
                 ))}
             </div>
         </div>
-    );
-};
+    )
+}