+++ SharePointFramework/ProjectExtensions/src/extensions/projectSetup/IProjectSetupData.ts
@@ -1,6 +1,6 @@
-import { ListContentConfig, ProjectTemplate, ProjectExtension } from '../../models/index';
-import { IHubSite } from 'sp-hubsite-service';
-import { ITemplateSelectDialogState } from '../../components/TemplateSelectDialog/ITemplateSelectDialogState';
+import { ListContentConfig, ProjectTemplate, ProjectExtension } from '../../models/index'
+import { IHubSite } from 'sp-hubsite-service'
+import { ITemplateSelectDialogState } from '../../components/TemplateSelectDialog/ITemplateSelectDialogState'
 
 export interface IProjectSetupData extends ITemplateSelectDialogState {
     /**