+++ SharePointFramework/ProjectExtensions/src/tasks/PlannerConfiguration/IPlannerBucket.ts
@@ -1,5 +1,0 @@
-export interface IPlannerBucket {
-    id: string;
-    name: string;
-    planId: string;
-}