+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Show only benefits (gevinster) that are toggled as "Show at portfolio level" #268
+- Support for attachments on provisioned planner tasks #274
 
 #### 1.2.2 - 2020-06-24